assistify-api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "assistify-api"
dynamic = ["version"]
description = "A robust FastAPI application that integrates with OpenAI's API to facilitate seamless conversation capabilities through a RESTful interface."
readme = "README.md"
authors = [{ name = "Justin Beall", email = "jus.beall@gmail.com" }]
requires-python = ">=3.12"
dependencies = [
    "ai-assistant-manager",
    "fastapi",
    "loguru",
    "google-auth",
    "openai",
    "python-dotenv",
    "uvicorn",
    "requests",
    "authlib",
]

[tool.hatch.version]
path = "setup.cfg"
pattern = "version = (?P<version>\\S+)"

[tool.hatch.build.targets.sdist]
include = ["/src"]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.envs.default]
type = "virtual"
path = ".venv"
dependencies = ["pyright", "pytest-cov", "pytest"]

[tool.hatch.envs.default.scripts]
dev = "python server.py"
start-app = "uvicorn assistify_api.app.api:api --host 0.0.0.0 --port 8000"
test = "pytest --cache-clear --cov --cov-report lcov --cov-report term"

[tool.hatch.envs.hatch-static-analysis]
config-path = "ruff_defaults.toml"

[tool.ruff]
extend = "ruff_defaults.toml"

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "parents"

[tool.pytest.ini_options]
markers = "integration: an integration test that hits external uncontrolled systems"
